+++ cmd/root.go
@@ -122,6 +122,7 @@ func Prepare() *cobra.Command {
 	rootCmd.AddCommand(convertCmd())
 	rootCmd.AddCommand(baselineCmd())
 	rootCmd.AddCommand(materializeCmd())
+	rootCmd.AddCommand(stampCmd())
 	rootCmd.AddCommand(validateCmd)
 
 	return rootCmd